kgo-verifier: do looping on consumers
This replaces a bash loop in tests.  To enable controlling
it, there is a new last_pass HTTP endpoint to ask it to
stop after it's current read, and a new 'active' element
in the status object to let the test know when this
worker has dropped out and its final status may be read.