kgo-verifier: make seq_read_worker more verbose
Historically, I kept these logs very terse for the benefit
of interactive use and because in earlier ducktape code
we took the stdout/stderr of the process over an SSH connection
back to the test runner.

Now that the services keep local log files that are hauled
back at the end of a test, we can afford to be much more verbose.

We still don't log the vector of offsets at info level on every poll call,
because that would be a huge log for ManyPartitionsTest.  Tests should
tune the logging up/down as needed: e.g. tests that operate on smaller
partition counts & smaller volumes of data can safely enable debug
logging.