kgo-verifier: track max consumed offset
This commit extends the status returned by the kgo-verifier workers
to contain the high watermarks for each partition (i.e. the max offset
consumed/produced).

A number of our tests that use the kgo-verifier consumer check
the number of valid reads in an attempt to ensure that all produced
offsets have been consumed at least once. Since the consumer follows a
sea-saw pattern (go back to the start offset when the end is reached),
this check is not necessarily correct. The new fields introduced in the
status can be used to improve those tests.