kgo-verifier: return status alongside lock
e02458e added a map to the various status structs returned by workers.
This turned out to be problematic as JSON serialisation could race with
mutation of the struct. This patch changes the GetStatus method to
return a tuple made out of the status and its lock. This approach was
taken because the status types differ and their locks are in different
places.