kgo-verifier: add configurable timestamp step
Retention tests would benefit from spacing out their timestamps more, so
this patch introduces a new flag for the producer to control the
timestamp step. It's only used when fake timestamps are enabled.

I've also done a little refactor and removed the fakeTimestampMs from
the producer worker. The value from the config struct is used instead.