kgo: run one extra pass after /last_pass request
This is being invoked by stress tests which run consumers in a loop but
want to ensure at the end that all the data was read.

kgo first requests high topic watermark, then starts routines that read
up to that watermark. If the last run begins before all the data is
written, then `/last_pass` is called, it will exit without having a
chance to read all the data.