verifier: do not crash with index out of range for invalid data
kgo expects to consume messages written by kgo itself. These messages
usually contain data (in headers) which is used for validation.

If unrelated data is produced to a partition it crashes because the
header it expects does not exist.

This commit avoids the crash and categorizes either categorizes the read
as out of scope or crashes if it should have been a valid read (running
with a kgo produced valid_offsets json).