crash instantly on data loss errors
We (nvartolomei and bash) found that GroupReadWorker wouldn't report
data loss. The reason is that we didn't fail the test instantly when
data loss is detected instead we relied on monotonicity validation to
fail. I.e. we would crash when we detected that we would consume an
earlier offset compared to what we consumed before.

In the GroupReadWorker we reset the monotonicity validation state on any
errors because the next attempt at consuming is almost certain to read
the second time offsets already consumed. As a result we wouldn't catch
monotonicity issues.

In retrospect, it would have been better to fail instantly when franz-go
detects data loss. So this commit achieves exactly that.