`verifier`: fix `lastConsumableOffsets`
This didn't account for tombstones being produced for the same key
later in the log.

Fix the logic by adding a new `latestKoByPartition` map which
tracks the last offset for a given key. By using it in combination
with `latestKvByPartition`, we can conclude which non-tombstone record/key
has the highest offset for a given partition.