kgo-repeater: listen for SIGTERM too
We listen for SIGINT and do a graceful shutdown, but we should do the
same thing on SIGTERM too, as that's what systemd will send a
service, and we run kgo as a service (and anyway it just makes sense
to do graceful shutdown on SIGTERM, that's kind of what it's there
for).

The main reason to do graceful shutdown is to commit any cached offsets
and (especially) to leave the consumer group explicitly, otherwise the
CG will stay in the Stable state until heartbeat timeout, which causes
problems for re-joining members.